pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blue-raven-cli"
version = "0.1.0"
description = "CLI for Featherweight Blue Raven over USB: download, configure, test"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "Cameron K. Brooks", email = "cambrooks3393@gmail.com" }]
license = { text = "AGPLv3" }
dependencies = [
  "pyserial>=3.5",
  "typer[all]>=0.9.0",
  "rich>=13.0.0",
]

[project.scripts]
braven = "braven.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["braven*"]
